+++ lib/application/repo/state/query_histories.dart
