+++ lib/presentation/page/repo/avatar_preview_page.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../application/repo/provider/current_repo.dart';
+import '../../../application/repo/state/current_repo.dart';
 import '../../../domain/repository/repo/entity/repo.dart';
 import '../../../util/logger.dart';
 import '../../component/async_value_handler.dart';
