+++ lib/presentation/page/repo/component/repo_detail_view.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../application/repo/provider/current_repo.dart';
+import '../../../../application/repo/state/current_repo.dart';
 import '../../../../application/url_launcher/url_launcher_service.dart';
 import '../../../../domain/repository/repo/entity/repo.dart';
 import '../../../../util/logger.dart';
