+++ lib/presentation/page/repo/component/search_repos_query.dart
@@ -5,8 +5,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../application/repo/provider/search_repos_query.dart';
 import '../../../../application/repo/search_repos_service.dart';
+import '../../../../application/repo/state/search_repos_query.dart';
 import '../../../../util/localization/strings.g.dart';
 import '../../../../util/logger.dart';
 import '../../../../util/routing/router.dart';
