+++ lib/presentation/page/repo/repo_view_page.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../application/repo/provider/current_repo.dart';
+import '../../../application/repo/state/current_repo.dart';
 import '../../../domain/repository/repo/entity/repo.dart';
 import '../../component/async_value_handler.dart';
 import '../../component/search_app_bar.dart';
